+++ src/kraken.rs
@@ -1,29 +1,21 @@
-use std::fmt::Display;
+use std::{convert::TryFrom, fmt::Display};
 
-use crate::taxonomy::Rank;
+use tracing::instrument;
+
+use crate::{errors::SpideogError, parser::parse_ident_organism_name, taxonomy::Rank};
 
 pub type ReportRecord = (String, u64, u64, Rank, u64, String);
-
-// pub struct KrakenReportRecord {
-//     pub percentage: String,
-//     pub fragments_clade: u64,
-//     pub fragments_taxon: u64,
-//     pub rank: TaxonomyRank,
-//     pub taxid: u64,
-//     pub indented_name: String,
-// }
-
 pub type Indent = usize;
 
 #[derive(Clone, PartialEq, PartialOrd, Debug, Ord, Eq, Hash, Deserialize)]
-pub struct Organism {
+pub struct Taxon {
     #[serde(rename = "taxonomy_lvl")]
     pub taxonomy_level: Rank,
     pub name: String,
     pub taxonomy_id: u64,
 }
 
-impl Display for Organism {
+impl Display for Taxon {
     fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
         write!(
             f,
@@ -33,9 +25,44 @@ impl Display for Organism {
     }
 }
 
-#[derive(Clone, PartialEq, PartialOrd, Debug, Serialize, Deserialize)]
+impl TryFrom<ReportRecord> for Taxon {
+    type Error = SpideogError;
+
+    #[instrument]
+    fn try_from(value: ReportRecord) -> Result<Self, Self::Error> {
+        let (_, (_, name)) = parse_ident_organism_name(value.5.as_bytes())
+            .map_err(|_e| SpideogError::KrakenIndentParsing)?;
+
+        let taxon = Self {
+            taxonomy_level: value.3,
+            name: String::from_utf8_lossy(name).trim().to_string(),
+            taxonomy_id: value.4,
+        };
+
+        Ok(taxon)
+    }
+}
+
+#[derive(Clone, PartialEq, PartialOrd, Debug, Serialize, Deserialize, Default)]
 pub struct Fragments {
-    pub percentage: f64,
-    pub count_clade: u64,
-    pub count_taxon: u64,
+    pub clade_percentage: f64,
+    pub clade_count_reads: u64,
+    pub taxon_count_reads: u64,
+}
+
+impl TryFrom<ReportRecord> for Fragments {
+    type Error = SpideogError;
+
+    #[instrument]
+    fn try_from(value: ReportRecord) -> Result<Self, Self::Error> {
+        let percentage = value.0.parse::<f64>().map_err(|_e| SpideogError::Other)?;
+
+        let fragments = Self {
+            clade_percentage: percentage,
+            clade_count_reads: value.1,
+            taxon_count_reads: value.1,
+        };
+
+        Ok(fragments)
+    }
 }