+++ src/subcommands.rs
@@ -1,2 +1,4 @@
+mod combine_abundance;
+mod combine_phylo;
+mod convert_abundance;
 mod convert_phylo;
-mod merge_phylo;