+++ CHANGELOG.md
@@ -8,6 +8,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased] - ReleaseDate
 
+### Added
+
+- command `convert-abundance`
+- command `combine-abundance`
+
+
+### Modified
+
+- fixed README typos
+
 ## [0.1.2] - 2020-10-29
 
 ### Added